requirements.txt
openenv-core
fastapi
uvicorn[standard]
fastmcp
pydantic>=2.0
openai
sqlparse
requests
google-genai